fix(runtime): fail closed on managed skill lock timeout
